merge...

+++ src/gKit/draw.cpp
@@ -7,7 +7,7 @@ void draw( Mesh& m, const Transform& model, const Orbiter& camera, GLuint textur
 {
     // recupere les transformations
     Transform view= orbiter_view_transform(camera);
-    Transform projection= orbiter_projection_transform(camera, window_width(), window_height(), 45);
+    Transform projection= orbiter_projection_transform(camera, (float) window_width(), (float) window_height(), 45);
     
     // affiche l'objet
     draw(m, model, view, projection, texture);
